% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distfuncs.R
\name{dmlog}
\alias{dmlog}
\alias{pmlog}
\alias{qmlog}
\alias{rmlog}
\title{The metalog distribution}
\usage{
dmlog(x, ml, log = FALSE)

pmlog(q, ml, lower.tail = TRUE, log.p = FALSE)

qmlog(p, ml, lower.tail = TRUE, log.p = FALSE)

rmlog(n, ml)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{ml}{Specification of the metalog distribution. Normally this is
obtained via a call to \code{\link{mlog}}.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE, probabilities are \eqn{P[X \le x]},
otherwise \eqn{P[X > x]}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
\code{dmlog} gives the density, \code{pmlog} gives the distribution
function, \code{qmlog} gives the quantile function, and \code{rmlog}
generates random deviates.

The length of the result is determined by \code{n} for \code{rmlog}, and by
the length of the first argument (\code{x}, \code{q}, or \code{p}) for
the other functions.
}
\description{
Density, distribution function, quantile function, and random generation for
the metalog distribution specified by \code{ml}.
}
\examples{
ml = mlog(quakes$lat);

x = seq(min(quakes$lat), max(quakes$lat), length.out = 100);
d = dmlog(x, ml);
p = pmlog(x, ml);

plot(x, d);
hist(rmlog(1000, ml), breaks = 100);

plot(x, p);
plot(x, qmlog(p, ml));

}
\references{
T.W. Keelin (2016) The metalog distributions. \emph{Decision
Analysis} 13: 243-277. \url{https://doi.org/10.1287/deca.2016.0338}.
}
