% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalog.R
\name{mlog}
\alias{mlog}
\title{Build a metalog distribution}
\usage{
mlog(
  x,
  y = "data",
  bounds = c(-Inf, Inf),
  n = NULL,
  mean = NULL,
  method = "any",
  refine = TRUE,
  check_validity = TRUE
)
}
\arguments{
\item{x}{A numeric vector of length 2 or greater. Interpretation depends on
the value of \code{y}.}

\item{y}{One of:
\itemize{
 \item{A numeric vector the same length as \code{x}: treated as probabilities
 corresponding to each value in \code{x}, defining the cumulative
 distribution function to be fitted. Both \code{x} and \code{y} must be
 specified in increasing order.}
 \item{A single number: treated as a confidence level, e.g. with 0.95 meaning
 a 95\% confidence interval. In this case, \code{x} should be either 2 or 3
 elements long, specifying either \code{c(ci_lower, ci_upper)} or
 \code{c(ci_lower, median, ci_upper)}.}
 \item{\code{"data"}: values in \code{x} are treated as empirical data
 (i.e. draws from an underlying distribution) to which the metalog
 distribution is fitted.}
 \item{\code{"bin"}: like \code{"data"}, but the data are binned first using
 base R's \code{hist} function. This is faster for very large data sets, but
 using \code{"data"} seems to generally give better results.}
 \item{\code{"coeff"}: the \code{x}s directly specify the coefficients
 \ifelse{html}{\out{<i>a<sub>i</sub></i>}}{\eqn{a_i}} of the metalog
 distribution.}
}
Character values for \code{y} can be abbreviated.}

\item{bounds}{Lower and upper bounds for the distribution. Can be one of:
\itemize{
 \item{\code{c(-Inf, Inf)}: Unbounded distribution, with support over the
 entire real line. The default.}
 \item{\code{c(bl, Inf)}: Semi-bounded distribution, with support over
 (\code{bl}, \eqn{\infty}).}
 \item{\code{c(-Inf, bu)}: Semi-bounded distribution, with support over
 (\eqn{-\infty}, \code{bu}).}
 \item{\code{c(bl, bu)}: Bounded distribution, with support over
 (\code{bl}, \code{bu}).}
}}

\item{n}{Number of terms in the metalog distribution. Must be 2 or greater. A
metalog distribution with \eqn{n} terms can, in principle, fit \eqn{n}
quantiles exactly.}

\item{mean}{If non-NULL, the mean of the distribution to be fitted. Only
works for bounded distributions.}

\item{method}{Which method to use in fitting the metalog distribution (when
\code{y} is something other than \code{"coeff"}). One of:
\itemize{
 \item{\code{"any"}: Try OLS first; if this fails to identify a feasible
 distribution, then try LP.}
 \item{\code{"ols"}: Ordinary least squares. This is the faster of the two
 methods, but will fail if the resulting metalog distribution is not
 feasible (see details).}
 \item{\code{"lp"}: Linear programming, using the R package \code{lpSolve}.
 This attempts to find a feasible metalog distribution that approximates the
 provided data as closely as possible.}
 \item{\code{"nl"}: Nonlinear optimization, using the R package \code{nloptr}.
 This attempts to find a feasible metalog distribution that approximates the
 provided data as closely as possible.}
 \item{\code{"cvx"}: Convex optimization, using the R package \code{CVXR}.
 This attempts to find a feasible metalog distribution that approximates the
 provided data as closely as possible.}
}
Character values for \code{method} can be abbreviated.}

\item{refine}{Whether to refine the cached version of the metalog quantile
function and PDF. This is enabled in order to slightly speed up calls to
\link{pmlog} and \link{dmlog}, although testing has revealed that this
sometimes has the opposite effect, and this parameter may be removed in the
future.}
}
\value{
An \code{mlog} object with elements:
\item{a}{The coefficients \ifelse{html}{\out{<i>a<sub>i</sub></i>}}{\eqn{a_i}}
of the fitted metalog distribution.}
\item{b}{The bounds of the fitted metalog distribution; same as the parameter
\code{bounds} supplied to the function.}
\item{cache}{A \code{data.frame} containing a cached version of the metalog
quantile function \eqn{M} and PDF \eqn{m}, as specified in eqs. 6 and 9 in
Keelin (2016).}
\item{method}{A character string describing the method that was actually
used to determine the coefficients \ifelse{html}{\out{<i>a<sub>i</sub></i>}}{\eqn{a_i}},
either \code{"ordinary least squares"}, \code{"linear programming"}, or
\code{"specified coefficients"}.}
}
\description{
Fits a metalog distribution to specified quantiles or data.
}
\details{
Not all sets of coefficients \ifelse{html}{\out{<i>a<sub>i</sub></i>}}{\eqn{a_i}}
result in a feasible metalog distribution. Metalog distributions are designed
to be very flexible so that they can approximate a wide variety of different
probability distributions. But the cost of this flexibility is that it is
possible to specify a set of coefficients \ifelse{html}{\out{<i>a<sub>i</sub></i>}}{\eqn{a_i}}
resulting in a PDF that goes negative in some areas, or equivalently, with a
CDF that is decreasing in some areas. This makes for an invalid probability
distribution, which means that the coefficients are not "feasible" in the
terminology of Keelin (2016).

In particular, the OLS (ordinary least squares) method of fitting metalog
distributions, which attempts to fit the specified quantiles as closely as
possible (or exactly in the case that the number of terms \code{n} is
equal to the number of specified quantiles), is not guaranteed to result in a
feasible distribution, though it often does, especially when the number of
terms \code{n} is lower. The LP (linear programming) method tries to ensure
that the resulting distribution is feasible, but it does this by compromising
on the exactness of the distribution's fit to the specified data, and the
success of this method cannot be guaranteed in general.

See Keelin (2016) for more details on feasibility.
}
\examples{
# UNBOUNDED DISTRIBUTIONS

# These are all equivalent:
ml1 = mlog(c(20, 30), c(0.05, 0.95))
ml2 = mlog(c(20, 25, 30), c(0.05, 0.5, 0.95))
ml3 = mlog(c(20, 30), 0.9)
ml4 = mlog(c(20, 25, 30), 0.9)

# A skewed distribution:
ml5 = mlog(c(20, 28, 30), 0.9)

# An empirically fitted distribution:
ml6 = mlog(quakes$lat)

}
\references{
T.W. Keelin (2016) The metalog distributions. \emph{Decision
Analysis} 13: 243-277. \url{https://doi.org/10.1287/deca.2016.0338}.
}
