% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{moment}
\alias{moment}
\title{Mean, variance, and other moments of a metalog distribution}
\usage{
moment(ml, order = 1, central = FALSE, absolute = FALSE)
}
\arguments{
\item{order}{Which moment to calculate.}

\item{central}{\code{TRUE} for moment about the mean (central moment),
\code{FALSE} for moment about 0.}

\item{absolute}{\code{TRUE} to calculate absolute moments.}
}
\description{
Calculate the moments of a metalog distribution \code{ml}. This function
allows calculating the mean, variance, and other quantities of potential
interest.
}
\details{
The mean of a distribution is its first noncentral moment and the variance
\ifelse{html}{\eqn{\sigma}\out{<sup>2</sup>}}{\eqn{\sigma^2}} is the second
central moment. Standardly, the skewness is defined as the third standardized
moment and the kurtosis as the fourth standardized moment, where the
\emph{n}th standardized moment is the \emph{n}th central moment divided by
\ifelse{html}{\eqn{\sigma}\out{<sup><i>n</i></sup>}}{\eqn{\sigma^n}}.

In other words:
\itemize{
 \item{Mean: \code{moment(ml)}}
 \item{Variance: \code{moment(ml, 2, TRUE)}}
 \item{Skewness: \code{moment(ml, 3, TRUE) / moment(ml, 2, TRUE)^(3/2)}}
 \item{Kurtosis: \code{moment(ml, 4, TRUE) / moment(ml, 2, TRUE)^2}}
}

Mathematically, this function calculates the integral over \eqn{y} from 0 to
1 of \ifelse{html}{\eqn{f(M(y)-c)}\out{<sup><i>k</i></sup>}}{\eqn{f(M(y)-c)^k}},
where: \eqn{M(y)} is the quantile function of the distribution; \eqn{k} is
\code{order}; \eqn{c} is equal to 0 if \code{central == FALSE} and is equal
to the mean of the distribution if \code{central == TRUE}; and \eqn{f(x)=x}
if \code{absolute == FALSE} and \eqn{f(x)=|x|} if \code{absolute == TRUE}.
}
