% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xtabulate}
\alias{xtabulate}
\title{Tabulation for numeric vectors}
\usage{
xtabulate(y)
}
\arguments{
\item{y}{a numeric vector.}
}
\value{
A list with two components: \code{x}, the sorted unique values in
\code{y}, and \code{n}, the count associated with each value in \code{x}.
}
\description{
Base-R's \code{\link[base]{tabulate}} will not tabulate numeric values; this
function does.
}
